refactor(classes): moving controller in a package
