+++ src/main/java/fr/coding/sfq/controllers/MainController.java
@@ -1,10 +1,9 @@
-package fr.coding.sfq;
+package fr.coding.sfq.controllers;
 
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.control.Button;
-import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.VBox;
 
 import java.io.IOException;
