+++ src/main/resources/fr/coding/sfq/HomePage.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.*?>
 
 <?import javafx.scene.text.Text?>
-<VBox xmlns:fx="http://javafx.com/fxml" fx:controller="fr.coding.sfq.HomeController" spacing="20" alignment="CENTER">
+<VBox xmlns:fx="http://javafx.com/fxml" fx:controller="fr.coding.sfq.controllers.HomeController" spacing="20" alignment="CENTER">
     <Label text="Bienvenue au Restaurant" style="-fx-font-size: 20px;"/>
     <Text>HomePage</Text>
     <fx:include source="DashboardView.fxml"/>
