+++ src/main/resources/fr/coding/sfq/Timer.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.layout.*?>
 
 <HBox xmlns:fx="http://javafx.com/fxml"
-      fx:controller="fr.coding.sfq.TimerController"
+      fx:controller="fr.coding.sfq.controllers.TimerController"
       alignment="CENTER_RIGHT" spacing="10" style="-fx-padding: 10;">
     <Label text="Temps restant :" />
     <Label fx:id="timerLabel" style="-fx-font-size: 16px; -fx-font-weight: bold;" />
