+++ src/main/resources/fr/coding/sfq/mainView.fxml
@@ -2,7 +2,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<BorderPane xmlns:fx="http://javafx.com/fxml" fx:controller="fr.coding.sfq.MainController">
+<BorderPane xmlns:fx="http://javafx.com/fxml" fx:controller="fr.coding.sfq.controllers.MainController">
     <top>
         <HBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml">
             <Label text="Restaurant Manager" style="-fx-font-size: 24px; -fx-font-weight: bold;" BorderPane.alignment="CENTER"/>
